#34 Move Confidence to gear package
