+++ retro-crawler-core/src/main/java/com/retrocrawler/core/gear/GearResolver.java
@@ -9,7 +9,6 @@ import java.util.function.Function;
 
 import com.retrocrawler.core.archive.clues.Artifact;
 import com.retrocrawler.core.archive.clues.Clue;
-import com.retrocrawler.core.archive.clues.Confidence;
 import com.retrocrawler.core.gear.injector.GearSpecialist;
 import com.retrocrawler.core.gear.matcher.GearMatcher;
 import com.retrocrawler.core.gear.parser.ParseContext;
