+++ retro-crawler-core/src/main/java/com/retrocrawler/core/gear/matcher/GearMatcher.java
@@ -1,6 +1,6 @@
 package com.retrocrawler.core.gear.matcher;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 import com.retrocrawler.core.gear.GearContext;
 
 public interface GearMatcher {
