+++ retro-crawler-core/src/test/java/com/retrocrawler/core/ContextualFactResolutionTest.java
@@ -26,8 +26,8 @@ import com.retrocrawler.core.archive.ArchiveId;
 import com.retrocrawler.core.archive.InMemoryRepository;
 import com.retrocrawler.core.archive.ReindexScope;
 import com.retrocrawler.core.archive.clues.Clue;
-import com.retrocrawler.core.archive.clues.Confidence;
 import com.retrocrawler.core.archive.clues.FolderNameClueFinder;
+import com.retrocrawler.core.gear.Confidence;
 import com.retrocrawler.core.gear.Fact;
 import com.retrocrawler.core.gear.GearContext;
 import com.retrocrawler.core.gear.RatedFact;
