+++ retro-crawler-core/src/test/java/com/retrocrawler/core/gear/RatedFactTest.java
@@ -6,8 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
-
 class RatedFactTest {
 
 	private enum Bus {
