+++ retro-crawler-core/src/test/java/com/retrocrawler/core/gear/parser/PathParserTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 
 import com.retrocrawler.core.Configuration;
 import com.retrocrawler.core.archive.Node;
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 import com.retrocrawler.core.gear.RatedFact;
 
 class PathParserTest {
