+++ retro-crawler-model/src/test/java/com/retrocrawler/model/identifier/NintendoGameBoyCartridgeCodeTest.java
@@ -8,7 +8,7 @@ import java.util.OptionalInt;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 
 class NintendoGameBoyCartridgeCodeTest {
