+++ retro-crawler-model/src/test/java/com/retrocrawler/model/identifier/PlayStationPortableDiscIdTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 
 class PlayStationPortableDiscIdTest {
