+++ retro-crawler-model/src/test/java/com/retrocrawler/model/measurement/CapacitySetTest.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 import com.retrocrawler.model.measurement.DataCapacity.Unit;
 
 class CapacitySetTest {
