+++ retro-crawler-model/src/test/java/com/retrocrawler/model/packaging/PackagingParsersTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 
 class PackagingParsersTest {
