+++ retro-crawler-model/src/test/java/com/retrocrawler/model/temporal/DateMarkingParserTest.java
@@ -13,7 +13,7 @@ import java.time.ZoneOffset;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 
 class DateMarkingParserTest {
