+++ retro-crawler-model/src/test/java/com/retrocrawler/model/temporal/YearParserTest.java
@@ -10,7 +10,7 @@ import java.time.ZoneOffset;
 
 import org.junit.jupiter.api.Test;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 
 class YearParserTest {
