+++ retro-crawler-mycollection/src/main/java/com/retrocrawler/mycollection/matchers/HardDiskDriveMatcher.java
@@ -1,6 +1,6 @@
 package com.retrocrawler.mycollection.matchers;
 
-import com.retrocrawler.core.archive.clues.Confidence;
+import com.retrocrawler.core.gear.Confidence;
 import com.retrocrawler.core.gear.GearContext;
 import com.retrocrawler.core.gear.matcher.GearMatcher;
 import com.retrocrawler.mycollection.AttributeNames;
